chore: renaming MTV keywork to Mtv
